+++ src/nonebot_plugin_resolver2/utils.py
@@ -1,6 +1,8 @@
 import asyncio
+import hashlib
 from pathlib import Path
 import re
+from urllib.parse import urlparse
 
 from nonebot import logger
 
@@ -40,3 +42,143 @@ async def exec_ffmpeg_cmd(cmd: list[str]) -> None:
     if return_code != 0:
         error_msg = stderr.decode().strip()
         raise RuntimeError(f"ffmpeg 执行失败: {error_msg}")
+
+
+async def merge_av(
+    *,
+    v_path: Path,
+    a_path: Path,
+    output_path: Path,
+) -> None:
+    """合并视频和音频
+
+    Args:
+        v_path (Path): 视频文件路径
+        a_path (Path): 音频文件路径
+        output_path (Path): 输出文件路径
+    """
+    logger.info(f"Merging {v_path.name} and {a_path.name} to {output_path.name}")
+
+    cmd = [
+        "ffmpeg",
+        "-y",
+        "-i",
+        str(v_path),
+        "-i",
+        str(a_path),
+        "-c",
+        "copy",
+        "-map",
+        "0:v:0",
+        "-map",
+        "1:a:0",
+        str(output_path),
+    ]
+
+    await exec_ffmpeg_cmd(cmd)
+    await asyncio.gather(safe_unlink(v_path), safe_unlink(a_path))
+    logger.success(f"Merged {output_path.name}, {fmt_size(output_path)}")
+
+
+async def merge_av_h264(
+    *,
+    v_path: Path,
+    a_path: Path,
+    output_path: Path,
+) -> None:
+    """合并视频和音频，并使用 H.264 编码
+
+    Args:
+        v_path (Path): 视频文件路径
+        a_path (Path): 音频文件路径
+        output_path (Path): 输出文件路径
+    """
+    logger.info(f"Merging {v_path.name} and {a_path.name} to {output_path.name} with H.264")
+
+    # 修改命令以确保视频使用 H.264 编码
+    cmd = [
+        "ffmpeg",
+        "-y",
+        "-i",
+        str(v_path),
+        "-i",
+        str(a_path),
+        "-c:v",
+        "libx264",  # 明确指定使用 H.264 编码
+        "-preset",
+        "medium",  # 编码速度和质量的平衡
+        "-crf",
+        "23",  # 质量因子，值越低质量越高
+        "-c:a",
+        "aac",  # 音频使用 AAC 编码
+        "-b:a",
+        "128k",  # 音频比特率
+        "-map",
+        "0:v:0",
+        "-map",
+        "1:a:0",
+        str(output_path),
+    ]
+
+    await exec_ffmpeg_cmd(cmd)
+    await asyncio.gather(safe_unlink(v_path), safe_unlink(a_path))
+    logger.success(f"Merged {output_path.name} with H.264, {fmt_size(output_path)}")
+
+
+async def encode_video_to_h264(video_path: Path) -> Path:
+    """将视频重新编码到 h264
+
+    Args:
+        video_path (Path): 视频路径
+
+    Returns:
+        Path: 编码后的视频路径
+    """
+    output_path = video_path.with_name(f"{video_path.stem}_h264{video_path.suffix}")
+    if output_path.exists():
+        return output_path
+    cmd = [
+        "ffmpeg",
+        "-y",
+        "-i",
+        str(video_path),
+        "-c:v",
+        "libx264",
+        "-preset",
+        "medium",
+        "-crf",
+        "23",
+        str(output_path),
+    ]
+    await exec_ffmpeg_cmd(cmd)
+    logger.success(f"视频重新编码为 H.264 成功: {output_path}, {fmt_size(output_path)}")
+    await safe_unlink(video_path)
+    return output_path
+
+
+def fmt_size(file_path: Path) -> str:
+    """格式化文件大小
+
+    Args:
+        video_path (Path): 视频路径
+    """
+    return f"大小: {file_path.stat().st_size / 1024 / 1024:.2f} MB"
+
+
+def generate_file_name(url: str, default_suffix: str = "") -> str:
+    """根据 url 生成文件名
+
+    Args:
+        url (str): url
+        default_suffix (str, optional): 默认后缀. Defaults to "".
+
+    Returns:
+        str: 文件名
+    """
+    # 根据 url 获取文件后缀
+    path = Path(urlparse(url).path)
+    suffix = path.suffix if path.suffix else default_suffix
+    # 获取 url 的 md5 值
+    url_hash = hashlib.md5(url.encode()).hexdigest()[:16]
+    file_name = f"{url_hash}{suffix}"
+    return file_name
